Move the salt loader ``log_handlers`` out of the internal logging package
