--- doc/ref/configuration/logging/handlers/salt.log.handlers.fluent_mod.rst
@@ -1,5 +1,0 @@
-============================
-salt.log.handlers.fluent_mod
-============================
-
-.. automodule:: salt.log.handlers.fluent_mod
