--- doc/ref/configuration/logging/handlers/salt.log.handlers.log4mongo_mod.rst
@@ -1,5 +1,0 @@
-===============================
-salt.log.handlers.log4mongo_mod
-===============================
-
-.. automodule:: salt.log.handlers.log4mongo_mod
