--- doc/ref/configuration/logging/handlers/salt.log.handlers.logstash_mod.rst
@@ -1,5 +1,0 @@
-==============================
-salt.log.handlers.logstash_mod
-==============================
-
-.. automodule:: salt.log.handlers.logstash_mod
