--- doc/ref/configuration/logging/handlers/salt.log.handlers.sentry_mod.rst
@@ -1,5 +1,0 @@
-============================
-salt.log.handlers.sentry_mod
-============================
-
-.. automodule:: salt.log.handlers.sentry_mod
