+++ doc/ref/configuration/logging/handlers/salt.log_handlers.fluent_mod.rst
@@ -1,0 +1,5 @@
+============================
+salt.log_handlers.fluent_mod
+============================
+
+.. automodule:: salt.log_handlers.fluent_mod
