+++ doc/ref/configuration/logging/handlers/salt.log_handlers.log4mongo_mod.rst
@@ -1,0 +1,5 @@
+===============================
+salt.log_handlers.log4mongo_mod
+===============================
+
+.. automodule:: salt.log_handlers.log4mongo_mod
