+++ doc/ref/configuration/logging/handlers/salt.log_handlers.logstash_mod.rst
@@ -1,0 +1,5 @@
+==============================
+salt.log_handlers.logstash_mod
+==============================
+
+.. automodule:: salt.log_handlers.logstash_mod
