+++ doc/ref/configuration/logging/handlers/salt.log_handlers.sentry_mod.rst
@@ -1,0 +1,5 @@
+============================
+salt.log_handlers.sentry_mod
+============================
+
+.. automodule:: salt.log_handlers.sentry_mod
