+++ salt/log_handlers/log4mongo_mod.py
