+++ salt/log_handlers/logstash_mod.py
